rename file to WebPageViewProvider
